feat(event-entry): FCFS 응모 영속화 및 관련 도메인 리팩터링
  - EventEntry 엔티티를 Instant 기반 팩토리/업데이트 메서드로 재구성
  - EventEntryService를 추가해 Kafka 이벤트 upsert 흐름을 정리
  - FCFS 전략에서 승인/중복/마감 케이스를 영속화하도록 위임
  - KafkaProducerDto 및 엔트리 요청 DTO의 ID 타입을 Long으로 통일

+++ common-messaging/src/main/java/com/axon/messaging/dto/KafkaProducerDto.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class KafkaProducerDto {
     private CampaignType campaignType;
-    private int eventId;
-    private int userId;
-    private int productId;
+    private Long eventId;
+    private Long userId;
+    private Long productId;
     private Long timestamp;
 }
