+++ core-service/src/main/java/com/axon/core_service/repository/EventRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
     List<Event> findAllByCampaign_Id(Long campaignId);
 }
