+++ entry-service/src/main/java/com/axon/entry_service/controller/EntryController.java
@@ -15,7 +15,7 @@ import java.time.Instant;
 @RestController
 @RequestMapping("/api/v1/entries")
 @RequiredArgsConstructor
-@CrossOrigin(origins = "http://localhost:8080") // core-service의 frontendからのリクエストを許可
+@CrossOrigin(origins = "http://localhost:8080")
 public class EntryController {
     private final Kafka_Producer producer;
     private final String axon_topic = "event";
@@ -24,7 +24,7 @@ public class EntryController {
     public ResponseEntity<Void> createEntry(@RequestBody EntryRequestDto requestDto) {
         log.info("요청 확인 {}", requestDto);
         // TODO: 추후 Spring Security를 통해 JWT 토큰에서 실제 userId를 추출해야 함
-        int userId = 1; // 임시로 하드코딩
+        Long userId = Long.valueOf(1); // 임시로 하드코딩
 
         long timestamp = Instant.now().toEpochMilli();
